processing-service/requirements.txt
# Only what we need for the POC
pika==1.3.1           # RabbitMQ client
python-dotenv==1.0.0  # For environment variables
docling[convert,chunk]==2.24.0

uuid
psycopg2-binary
pgvector
openai
numpy
tiktoken
# Remove these for now as we're not using them yet
# fastapi==0.68.1
# uvicorn==0.15.0
# python-multipart==0.0.5
# sentence-transformers==2.2.2  # This was causing the long build time
# pytest==6.2.5 